test CPU as well
